+++ spatial-rdd/src/main/java/edu/utdallas/cg/spatial_rdd/core/query/range/RangeQuery.java
@@ -60,7 +60,6 @@ public class RangeQuery implements Serializable {
    *
    * @param spatialRDD the spatial RDD
    * @param queryWindow the original query window
-   * @param considerBoundaryIntersection the consider boundary intersection
    * @param useIndex the use index
    * @return the java RDD
    * @throws Exception the exception
@@ -69,7 +68,6 @@ public class RangeQuery implements Serializable {
   public static <U extends Geometry, T extends Geometry> JavaRDD<T> spatialRangeQuery(
       SpatialRDD<T> spatialRDD,
       Envelope queryWindow,
-      boolean considerBoundaryIntersection,
       boolean useIndex) {
     Coordinate[] coordinates = new Coordinate[5];
     coordinates[0] = new Coordinate(queryWindow.getMinX(), queryWindow.getMinY());
@@ -79,6 +77,6 @@ public class RangeQuery implements Serializable {
     coordinates[4] = coordinates[0];
     GeometryFactory geometryFactory = new GeometryFactory();
     U queryGeometry = (U) geometryFactory.createPolygon(coordinates);
-    return spatialRangeQuery(spatialRDD, queryGeometry, considerBoundaryIntersection, useIndex);
+    return spatialRangeQuery(spatialRDD, queryGeometry, false, useIndex);
   }
 }
